+++ CHANGELOG.md
@@ -1,3 +1,17 @@
 # Changelog
 
 ## Table of Contents
+
+ - **Version 0**
+   - Version 0.1
+     - [*Version 0.1.0-alpha*](#version-010-alpha-24042020)
+
+### Version 0.1.0-alpha (24/04/2020)
+
+> <sup>Node.js v14.0.0</sup>
+
+ - <kbd>minor</kbd> Created logging module: **logger.js v1.0.0**.
+ - <kbd>minor</kbd> Created SSL module: **ssl.js v1.0.0**.
+ - <kbd>minor</kbd> Added dependency: **acme-client v3.3.1** from [npm](https://www.npmjs.com/package/acme-client).
+ - <kbd>minor</kbd> Created Port configuration: **port.json**.
+ - <kbd>minor</kbd> Created SSL configuration: **ssl.json**.